Move Galley notification endpoints
