+++ libs/wire-api-federation/wire-api-federation.cabal
@@ -22,6 +22,7 @@ library
     Wire.API.Federation.API.Cargohold
     Wire.API.Federation.API.Common
     Wire.API.Federation.API.Galley
+    Wire.API.Federation.API.Galley.Notifications
     Wire.API.Federation.BackendNotifications
     Wire.API.Federation.Client
     Wire.API.Federation.Component
